+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 * #235 Add ``custom-is-checked`` prop to table column, add it when you want a custom compare function when is checkable
 * #237 Fix emit blur event for input
 * #239 Fix table columns on v-for loop
-* #241 Fix remove of tabItem when it's bounded in array
+* #241 Fix remove tabItem when it's bounded to array
+* #243 Add ``default-sort-direction`` prop to table, add it when you want change default sort direction on the first click
 * Add ``multiple`` support to select
 * Add ``inline`` prop to dropdown, items are shown inline and trigger removed